refactor: 한칸 움직임을 Dot 내부 메서드로 구현
