+++ README.md
@@ -52,4 +52,4 @@
     0 1 2 3 4 5 6 7 8
 ```
 
-- [ ] 게임 진행 순서는 초나라부터 시작한다.
+- [x] 게임 진행 순서는 초나라부터 시작한다.
