+++ src/main/java/Application.java
@@ -1,11 +1,10 @@
-import game.Dot;
-import game.JanggiGame;
-import game.arrangement.OuterElephantStrategy;
-import piece.Dynasty;
+import janggiGame.JanggiGame;
+import janggiGame.arrangement.OuterElephantStrategy;
+import janggiGame.board.Dot;
+import janggiGame.piece.Dynasty;
+import java.util.List;
 import view.InputView;
 import view.OutputView;
-
-import java.util.List;
 
 public class Application {
