+++ src/main/java/janggiGame/piece/Pawn.java
@@ -1,14 +1,14 @@
-package piece;
+package janggiGame.piece;
 
-import game.Dot;
-
+import janggiGame.board.Dot;
 import java.util.List;
 import java.util.Map;
 
-public class Advisor extends Piece {
-    public static final String NAME = "사";
+public class Pawn extends Piece {
 
-    public Advisor(Dynasty dynasty) {
+    public static final String NAME = "병";
+
+    public Pawn(Dynasty dynasty) {
         super(dynasty);
     }
 
@@ -19,6 +19,14 @@ public class Advisor extends Piece {
 
         if (Math.abs(dx) + Math.abs(dy) != 1) {
             throw new UnsupportedOperationException("[ERROR] 병이 이동할 수 있는 목적지가 아닙니다.");
+        }
+
+        if (dynasty == Dynasty.HAN && dy > 0) {
+            throw new UnsupportedOperationException("[ERROR] 병은 뒤로 이동할 수 없습니다.");
+        }
+
+        if (dynasty == Dynasty.CHO && dy < 0) {
+            throw new UnsupportedOperationException("[ERROR] 병은 뒤로 이동할 수 없습니다.");
         }
 
         return List.of(); //ToDO 빈값 확인해주기!!
